refactor home feature watching history screen
