--- feature/home/presentation/src/main/java/com/sanaa/presentation/screen/mediaTabScreen/continueWatchingScreen/ContinueWatchingScreenEffect.kt
@@ -1,8 +1,0 @@
-package com.sanaa.presentation.screen.mediaTabScreen.continueWatchingScreen
-
-import com.sanaa.presentation.state.MediaTypeUi
-
-sealed class ContinueWatchingScreenEffect {
-    object NavigateBack : ContinueWatchingScreenEffect()
-    data class NavigateToMediaDetails(val id: Int, val mediaTypeUi: MediaTypeUi) : ContinueWatchingScreenEffect()
-}
