+++ feature/home/presentation/src/main/java/com/sanaa/presentation/screen/mediaTabScreen/topRatingScreen/TopRatedScreenEffect.kt
@@ -2,8 +2,8 @@ package com.sanaa.presentation.screen.mediaTabScreen.topRatingScreen
 
 import com.sanaa.presentation.state.MediaTypeUi
 
-sealed class TopRatedScreenEffect{
-    object NavigateBack : TopRatedScreenEffect()
-    data class NavigateToMediaDetails(val id: Int, val mediaTypeUi: MediaTypeUi) : TopRatedScreenEffect()
-    object NavigateToLogin : TopRatedScreenEffect()
+sealed interface TopRatedScreenEffect{
+    object NavigateBack : TopRatedScreenEffect
+    data class NavigateToMediaDetails(val id: Int, val mediaTypeUi: MediaTypeUi) : TopRatedScreenEffect
+    object NavigateToLogin : TopRatedScreenEffect
 }
