+++ feature/home/presentation/src/main/java/com/sanaa/presentation/screen/mediaTabScreen/watchingHistoryScreen/WatchingHistoryScreenUiState.kt
@@ -1,10 +1,10 @@
-package com.sanaa.presentation.screen.mediaTabScreen.continueWatchingScreen
+package com.sanaa.presentation.screen.mediaTabScreen.watchingHistoryScreen
 
 import com.sanaa.presentation.state.GenreUiState
 import com.sanaa.presentation.state.MediaItem
 import com.sanaa.presentation.state.MediaTypeUi
 
-data class ContinueWatchingMediaScreenUiState(
+data class WatchingHistoryScreenUiState(
     val selectedMediaTypeUi: MediaTypeUi = MediaTypeUi.MOVIE,
     val movieList: List<MediaItem> = emptyList(),
     val tvShowList: List<MediaItem> = emptyList(),
