+++ feature/home/presentation/src/main/java/com/sanaa/presentation/screen/trendingMediaScreen/TrendingMediaScreenEffect.kt
@@ -1,7 +1,7 @@
 package com.sanaa.presentation.screen.trendingMediaScreen
 
-sealed class TrendingMediaScreenEffect {
-    object NavigateBack : TrendingMediaScreenEffect()
-    data class NavigateToMediaDetails(val id: Int) : TrendingMediaScreenEffect()
-    object NavigateToLogin: TrendingMediaScreenEffect()
+sealed interface TrendingMediaScreenEffect {
+    object NavigateBack : TrendingMediaScreenEffect
+    data class NavigateToMediaDetails(val id: Int) : TrendingMediaScreenEffect
+    object NavigateToLogin: TrendingMediaScreenEffect
 }
